requirements.txt
psycopg2-binary==2.9.5
python-telegram-bot==13.15
geopy==2.3.0